Seasonal decomposition with X-13ARIMA-SEATS

Decomposes a time series via model based signal extraction using X-13ARIMA-SEATS.

SEATS(.data, formula, ...)

Arguments

.data
A tsibble.

formula
Decomposition specification.

...
Other arguments passed to seasonal::seas().

Specials

season
The season special is used to specify seasonal attributes of the decomposition.

season(period = NULL)

period

The periodic nature of the seasonality. X-13ARIMA-SEATS decompositions only seasonal patterns with periods of 1, 2, 4, 6 or 12.

References

Dagum, E. B., & Bianconcini, S. (2016) "Seasonal adjustment methods and real time trend-cycle estimation". Springer.

X-13ARIMA-SEATS Documentation from the seasonal package's website: http://www.seasonal.website/seasonal.html

Official X-13ARIMA-SEATS manual: https://www.census.gov/ts/x13as/docX13ASHTML.pdf

See also

Examples

tsibbledata::aus_production %>% feasts:::SEATS(Beer)

#> Model used in SEATS is different: (0 1 1)(0 1 1)

#> # A dable:                       218 x 6 [1Q]
#> # X-13ARIMA-SEATS Decomposition: Beer = trend * seasonal * irregular
#>    Quarter  Beer trend seasonal irregular season_adjust
#>      <qtr> <dbl> <dbl>    <dbl>     <dbl>         <dbl>
#>  1 1956 Q1   284  255.    1.05      1.06           271.
#>  2 1956 Q2   213  255.    0.860     0.970          248.
#>  3 1956 Q3   227  256.    0.912     0.973          249.
#>  4 1956 Q4   308  257.    1.19      1.01           258.
#>  5 1957 Q1   262  258.    1.04      0.976          251.
#>  6 1957 Q2   228  259.    0.863     1.02           264.
#>  7 1957 Q3   236  260.    0.915     0.992          258.
#>  8 1957 Q4   320  261.    1.19      1.03           268.
#>  9 1958 Q1   272  262.    1.04      1.00           262.
#> 10 1958 Q2   233  262.    0.864     1.03           270.
#> # … with 208 more rows