X11 seasonal decomposition

Applies a seasonal adjustment by an enhanced version of the methodology of the Census Bureau X-11 and X-11Q programs. The type of seasonal decomposition (additive or multiplicative) can be controlled using the type argument.

X11(.data, formula, type = c("additive", "multiplicative"), ...)

Arguments

.data
A tsibble.

formula
Decomposition specification.

type
The type of decomposition: additive or multiplicative

...
Other arguments passed to seasonal::seas().

Specials

season
The season special is used to specify seasonal attributes of the decomposition.

season(period = NULL)

period

The periodic nature of the seasonality. X11 decompositions only support monthly (

period = 12	) and quarterly (

period = 4	) seasonal patterns.

References

Dagum, E. B., & Bianconcini, S. (2016) "Seasonal adjustment methods and real time trend-cycle estimation". Springer.

X11 Documentation from the seasonal package's website: http://www.seasonal.website/seasonal.html#input

Official X-13ARIMA-SEATS manual: https://www.census.gov/ts/x13as/docX13ASHTML.pdf

See also

Examples

tsibbledata::aus_production %>% X11(Beer)

#> # A dable:           218 x 6 [1Q]
#> # X11 Decomposition: Beer = trend + seasonal + irregular
#>    Quarter  Beer trend seasonal irregular season_adjust
#>      <qtr> <dbl> <dbl>    <dbl>     <dbl>         <dbl>
#>  1 1956 Q1   284  272.     6.36      5.66          278.
#>  2 1956 Q2   213  255.   -35.7      -6.71          249.
#>  3 1956 Q3   227  249.   -23.7       1.62          251.
#>  4 1956 Q4   308  252.    53.0       2.56          255.
#>  5 1957 Q1   262  257.     6.69     -2.18          255.
#>  6 1957 Q2   228  261.   -36.4       3.92          264.
#>  7 1957 Q3   236  263.   -22.8      -4.00          259.
#>  8 1957 Q4   320  265.    52.3       2.25          268.
#>  9 1958 Q1   272  267.     7.11     -1.94          265.
#> 10 1958 Q2   233  266.   -37.3       4.00          270.
#> # … with 208 more rows