STL features

Computes a variety of measures extracted from an STL decomposition of the time series. This includes details about the strength of trend and seasonality.

feat_stl(x, .period, s.window = 11, ...)

Arguments

x

A vector to extract features from.

.period

The period of the seasonality.

s.window

The seasonal window of the data (passed to stats::stl())

...

Further arguments passed to stats::stl()

Value

A vector of numeric features from a STL decomposition.

See also