Seasonal decomposition with X-13ARIMA-SEATS

Decomposes a time series via model based signal extraction using X-13ARIMA-SEATS.

X_13ARIMA_SEATS(
  formula,
  ...,
  defaults = c("none"),
  na.action = seasonal::na.x13
)

Arguments

formula
Decomposition specification.

...
Other arguments passed to seasonal::seas().

.data
A tsibble.

Specials

season

The season special is used to specify seasonal adjustment method for the decomposition.

season(period = NULL)

period	The periodic nature of the seasonality. X-13ARIMA-SEATS decompositions only seasonal patterns with periods of 1, 2, 4, 6 or 12.

References

Dagum, E. B., & Bianconcini, S. (2016) "Seasonal adjustment methods and real time trend-cycle estimation". Springer.

X-13ARIMA-SEATS Documentation from the seasonal package's website: http://www.seasonal.website/seasonal.html

Official X-13ARIMA-SEATS manual: https://www.census.gov/ts/x13as/docX13ASHTML.pdf

See also

Examples