Phillips and Ouliaris Cointegration Test

Performs the Phillips and Ouliaris "Pu" and "Pz" cointegration test.

cointegration_phillips_ouliaris(x, ...)

Arguments

x

Matrix of data to be tested.

...

Additional arguments passed to urca::ca.po().

Value

An object of class ca.po.

Details

The test "Pz", compared to the test "Pu", has the advantage that it is invariant to the normalization of the cointegration vector, i.e. it does not matter which variable is on the left hand side of the equation. In case convergence problems are encountered by matrix inversion, one can pass a higher tolerance level via "tol=..." to the solve()-function.

References

Phillips, P.C.B. and Ouliaris, S. (1990), Asymptotic Properties of Residual Based Tests for Cointegration, Econometrica, Vol. 58, No. 1, 165–193.

See also

Author

Bernhard Pfaff

Examples

cointegration_phillips_ouliaris(cbind(mdeaths, fdeaths))
#>   phillips_ouliaris_stat phillips_ouliaris_pvalue 
#>                  58.1279                   0.0100